TR09-ExceptionsAndHATEOS

+++ sgvet/src/main/java/br/edu/iff/ccc/bsi/sgvet/services/FuncionarioService.java
@@ -1,11 +1,12 @@
 package br.edu.iff.ccc.bsi.sgvet.services;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import br.edu.iff.ccc.bsi.sgvet.entities.Funcionario;
 import br.edu.iff.ccc.bsi.sgvet.repository.FuncionarioRepository;
@@ -21,8 +22,11 @@ public class FuncionarioService {
 	}
 	
 	public Funcionario getById(Long id) {
-		Optional<Funcionario> funcionario = funcionarioRep.findById(id);
-		return funcionario.orElse(null);
+		return funcionarioRep.findById(id)
+			.orElseThrow(() -> new ResponseStatusException(
+					HttpStatus.NOT_FOUND,
+					"Funcionario com ID " + id + " nao encontrado"
+			));
 	}
 	
 	@Transactional
@@ -32,6 +36,12 @@ public class FuncionarioService {
 	
 	@Transactional
 	public void delete(Long id) {
+		if(!funcionarioRep.existsById(id)) {
+			throw new ResponseStatusException(
+					HttpStatus.NOT_FOUND,
+					"Funcionario com ID " + id + " nao encontrado."
+			);
+		}
 		funcionarioRep.deleteById(id);
 	}
 }
